If Kron P is initialized on the fly, it whitens the first gradient it sees
PSGD Kron whitening preconditioner may converge slow for the first a few thousands of iterations if the preconditioner is initialized to an identity matrix. 

This commit initializes the preconditioner to whiten the first gradient that PSGD sees along the smallest dim. Need to set preconditioner_init_scale to None to enable this initialize-on-the-fly feature.