More conservative auto init scale for Kron preconditioner
1, The optimal preconditioner initial scale a can be derived from 

min_a g^T (a * I ) g + v^T (I/a) v,  v ~ N(0, 1)

For Kron preconditioners, previously we auto initialize the scale of each Q independently. Now, we use the min initial scale as the scale of all Q's. This is safer and also introduces the least priors to the Hess or FIM estimation. 

2, Other minor updates.