simplified subspace iteration
With CPU emulated bfloat16, there is a probability about 0.006 than randn([]) returns zero! This could happen in other places too. That means that the vectors in the initial noise subspace has a good probability to be all zeros when the dim of matrix is small. The exact implementation needs patches to handle such corner cases. Hard to justify all the complexities. So just simplified it an intuitive implementation.