add choice dP=P^0.5 E P^1.5
This choice directly corresponds to the Newton-Schulz iteration for matrix square root. It is useful with single or higher precisions, but not good for half precision. 

Also now the code does not prevent one from trying to fit P directly with half precision, even though most like this is not good.