demo for torch.optim distributed training wrapping
One can wrap the functional APIs as a torch.optim.Optimizer as showing here. 

Note that matmul is not deterministic in GPU. Generally there is no bitwise match guarantee of model parameters and preconditioner states for this DDP wrapping demo. Occasional sync of those states may be required if it is a concern.